myproject/pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.1.0",
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.5.3",
    "django-ninja-extra>=0.30.9",
    "django-ninja-jwt>=5.4.4",
    "django-pandas>=0.6.7",
    "matplotlib>=3.10.8",
    "pandas>=3.0.0",
    "pillow>=12.1.0",
    "psycopg[binary]>=3.3.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-django>=4.11.1",
    "pytest-env>=1.2.0",
    "ruff>=0.14.14",
]

[tool.ruff]
# 1行の長さ（Djangoのデフォルトに合わせるなら88、少し広めなら100〜120）
line-length = 100

[tool.ruff.lint]
# 適用するルールの選択 (E: pycodestyle, F: Pyflakes, I: isort)
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.lint.isort]
# Djangoのプロジェクト名を指定して、インポート順を綺麗にする
known-first-party = ["myproject"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "myproject.settings"
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

# 👇 これが超重要！
# テストを実行する時だけ、環境変数を「ローカル用」に書き換えます
env = [
    "D:DATABASE_URL=postgres://postgres:password@localhost:5432/mydb",
    "D:SECRET_KEY=test-secret-key",
    "D:NINJA_SKIP_REGISTRY=yes",
]
